Fragments
In react, always we have a root components (a div for example) to wrap our child components that can be multiple, then in the browser that will be rendering with the root controller (our div), what will we do if we want to get rid of the root and show all multiple child with as siblings of all other componets in the hierarchy?. Fragments is the response.
Fragments hide our root component as that doesn't exists, so our child components will be show as root components in the hierarchy.
e.g. Instead to put a root as <div>, you call put all like <> [components] </> without any type of markup to hide our root.